+++ tests/test_api.cpp
@@ -6,6 +6,7 @@
 #include "operators/embedding_op.h"
 #include "operators/matmul_op.h"
 #include "runtime/device_buffer.h"
+#include "runtime/cuda_compat.h"
 #include "tokenizer.h"
 #include <vector>
 
@@ -33,12 +34,12 @@ vector<float> TestAPI::get_embeddings(vector<int> token_ids) {
   io::StagedReader staged_reader(1 << 20);
 
   const io::TensorMeta& embed_meta = embedding_loader.meta("model.embed_tokens.weight");
-  runtime::DeviceBuffer<float> embedding_table(embed_meta.num_elements);
+  runtime::DeviceBuffer<__nv_bfloat16> embedding_table(embed_meta.num_elements);
   staged_reader.upload_tensor(embedding_loader, "model.embed_tokens.weight",
                               embedding_table.data(), context.stream());
   context.synchronize();
 
-  EmbeddingOp embedding_op(runtime::make_device_tensor_view<const float>(
+  EmbeddingOp embedding_op(runtime::make_device_tensor_view<const __nv_bfloat16>(
       embedding_table.data(), {embed_meta.shape[0], embed_meta.shape[1]}));
 
   runtime::DeviceBuffer<float> output(token_ids.size() * EMBEDDING_DIM);
